Delete RequestBodyBuilder
